requirements.txt
fastapi
uvicorn
requests
pillow
torch
torchvision
timm
jinja2
python-multipart